requirements.txt
coverage==4.2
coveralls==1.1
docopt==0.6.2
py==1.4.31
pytest==3.0.4
requests==2.12.1
paho-mqtt==1.2
RPi.GPIO==0.6.3